+++ packages/@plugin/plugin-scss/index.cjs
@@ -1,9 +1,9 @@
 const PluginConfig = require("./generator/index.cjs");
 
-const pluginSass = (buildTool) => {
+const pluginScss = (buildTool) => {
   return PluginConfig[buildTool] ?? console.warn(`Unsupported build tool: ${buildTool}`);
 };
 
 module.exports = {
-  pluginSass,
+  pluginScss,
 };
