+++ packages/core/src/models/protocolGenerator/PluginToTemplateAPI.ts
@@ -65,12 +65,6 @@ class PluginToTemplateAPI extends ProtocolGeneratorAPI {
   // 匹配 createApp 语句
   private createAppRegex: RegExp = /const\s+app\s*=\s*createApp\s*\(\s*App\s*\)/;
 
-  private StyleReg: { [key: string]: RegExp } = {
-    css: /css$/i,
-    scss: /scss$/i, // 匹配以 .scss 结尾的文件
-    less: /less$/i, // 匹配以 .less 结尾的文件
-  };
-
   constructor(protocols, props) {
     super(protocols);
     this.protocols = protocols;
@@ -209,93 +203,20 @@ class PluginToTemplateAPI extends ProtocolGeneratorAPI {
    * 样式类插件协议
    * @param params
    */
-  PROCESS_STYLE_PLUGIN() {
+  PROCESS_STYLE_PLUGIN(params) {
+    let content = params.params.content;
     const plugins = this.props.preset.plugins;
+    const template = this.props.preset.template;
     const fileData: FileData = this.props.files.getFileData();
     try {
       for (const plugin in plugins) {
-        if (plugins.hasOwnProperty(plugin) && plugin === "sass") {
-          this.processStyleFiles("scss", fileData);
+        if (plugins.hasOwnProperty(plugin) && plugin === "scss") {
+          content.processStyleFiles("scss", fileData, template, content.processScss);
         }
       }
     } catch (error) {
       console.log(error);
     }
-  }
-
-  /**
-   * 处理 sass、less 文件函数
-   * @param plugin 插件名
-   * @param fileData 文件内容
-   * @returns
-   */
-  private processStyleFiles(plugin: string, fileData: FileData): FileData {
-    const regexps = this.StyleReg["css"];
-    for (let i = 0; i < fileData.children.length; i++) {
-      // 先寻找 src 文件夹
-      const dirName = path.basename(fileData.children[i].path);
-      if (dirName === "src") {
-        const srcFileData = fileData.children[i];
-        for (let j = 0; j < srcFileData.children.length; j++) {
-          const dirName = path.basename(srcFileData.children[j].path);
-          const extension = dirName.split(".").pop();
-          if (regexps.test(extension)) {
-            const child = fileData.children[i].children[j];
-            const fileDescribe = child.describe;
-            // 更新文件扩展名和路径
-            if (child.path && child.path.endsWith(`.${fileDescribe.fileExtension}`)) {
-              child.describe.fileExtension = plugin;
-              child.path = child.path.replace(child.path.match(regexps)[0], `.${plugin}`);
-            }
-            // 文件内容由回调函数处理
-            if (fileDescribe && typeof fileDescribe.fileContent === "string") {
-              fileData = this.contentCallback(fileData, plugin);
-            }
-          }
-        }
-      }
-    }
-
-    return fileData;
-  }
-
-  /**
-   * fileContent 回调函数，执行 style 对应的处理函数
-   * @param fileContent 文件内容
-   * @param plugin 插件名称
-   * @returns
-   */
-  private contentCallback(fileContent: FileData, plugin: string): FileData {
-    const methodName = `process${plugin.charAt(0).toUpperCase() + plugin.slice(1)}`;
-    if (typeof this[methodName] === "function") {
-      return this[methodName](fileContent);
-    } else {
-      console.log(`No processing method for: ${plugin}`);
-    }
-  }
-
-  // 处理 sass 的回调函数，
-  private processScss(fileData: FileData): FileData {
-    const template = this.props.preset.template;
-    for (let i = 0; i < fileData.children.length; i++) {
-      const dirName = path.basename(fileData.children[i].path);
-      if (dirName === "src") {
-        if (template === "vue") {
-          const vueData = fileData.children[i].children[0].describe;
-          vueData.fileContent = vueData.fileContent.replace(
-            `</script>\n\n<style scoped>\n@import "./index.css";\n</style>\n`,
-            `</script>\n\n<style scoped lang="scss">\n@import "./index.scss";\n</style>\n`,
-          );
-        } else if (template === "react") {
-          const reactData = fileData.children[i].children[0].describe;
-          reactData.fileContent = reactData.fileContent.replace(
-            `import "./index.css"`,
-            `import "./index.scss"`,
-          );
-        }
-      }
-    }
-    return fileData;
   }
 }
